More strategy3() to dqn.py Refine CLI for dqn.py, change param C from 15k -> 10k.
